requirements.txt
asgiref==3.8.1
certifi==2024.7.4
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==4.2.14
django-cors-headers==4.4.0
djangorestframework==3.15.2
ffmpeg==1.4
ffmpeg-python==0.2.0
fpdf==1.7.2
future==1.0.0
idna==3.7
joblib==1.4.2
nltk==3.8.1
pillow==10.4.0
pydub==0.25.1
regex==2024.7.24
reportlab==4.2.2
requests==2.32.3
SpeechRecognition==3.10.4
sqlparse==0.5.1
tqdm==4.66.4
typing-extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
